curiocity(m8): fix §3 dependency-boundary — relocate computeTurnMetrics to results/
stats/turn-metrics.ts imported from interaction/, crossing the §3 module floor
(stats must import only shared/ + results types). The reducer is pure over
results/schema types, so move it to results/turn-metrics.ts; curion, stats, and
the test now import from there. No behavior change.
